chatbot_2/ingestapi_v1/requirement.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
requests==2.32.3
python-dotenv==1.0.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.22
tiktoken==0.7.0
pypdf==4.3.1
docx2txt==0.8
python-multipart
pgvector==0.2.5


